--- server/backend/testdata.csv
@@ -1,122 +1,0 @@
-[
-  {
-    "type": "burn",
-    "schedules": [
-      {
-        "type": "weekday",
-        "value": "3"
-      },
-      {
-        "type": "month",
-        "value": "5"
-      },
-      {
-        "type": "none",
-        "value": ""
-      }
-    ]
-  },
-  {
-    "type": "bottole",
-    "schedules": [
-      {
-        "type": "biweek",
-        "value": "1-1"
-      },
-      {
-        "type": "weekday",
-        "value": "4"
-      },
-      {
-        "type": "none",
-        "value": ""
-      }
-    ]
-  }
-],
-[
-  {
-    "type": "burn",
-    "schedules": [
-      {
-        "type": "weekday",
-        "value": "0"
-      },
-      {
-        "type": "none",
-        "value": ""
-      },
-      {
-        "type": "none",
-        "value": ""
-      }
-    ]
-  },
-  {
-    "type": "bottole",
-    "schedules": [
-      {
-        "type": "none"
-      },
-      {
-        "type": "biweek",
-        "value": "0-2"
-      },
-      {
-        "type": "none",
-        "value": ""
-      }
-    ]
-  },
-  {
-    "type": "petbottle",
-    "schedules": [
-      {
-        "type": "biweek",
-        "value": "1-4"
-      },
-      {
-        "type": "month",
-        "value": "12"
-      },
-      {
-        "type": "none",
-        "value": ""
-      }
-    ]
-  },
-  {
-    "type": "resource",
-    "schedules": [
-      {
-        "type": "month",
-        "value": "4"
-      },
-      {
-        "type": "none",
-        "value": ""
-      },
-      {
-        "type": "none",
-        "value": ""
-      }
-    ]
-  },
-  {
-    "type": "coarse",
-    "schedules": [
-      {
-        "type": "weekday",
-        "value": "2"
-      },
-      {
-        "type": "biweek",
-        "value": "1-5"
-      },
-      {
-        "type": "none",
-        "value": ""
-      }
-    ]
-  }
-]
