+++ server/backend/utility.js
@@ -1,3 +1,5 @@
+const moment = require('moment-timezone');
+
 exports.create_id = ()=>{
     let uuid = '', i, random;
     for (i = 0; i < 32; i++) {
@@ -14,22 +16,20 @@ exports.create_id = ()=>{
 /**
     params:offset 今週=0,来週=1
 **/
-const calculateStartDate = (offset) => {
-    const JSTOffset = 60 * 9 * 60 * 1000; // JST時間を求めるためのオフセット
-    var localdt = new Date(); // 実行サーバのローカル時間
-    var jsttime = localdt.getTime() + (localdt.getTimezoneOffset() * 60 * 1000) + JSTOffset;
-    // 開始日を日曜日にセットする
-    var dt = new Date(jsttime);
-    dt.setDate(dt.getDate() - dt.getDay() + (7 * offset));
+const calculateStartDate = (offset, tz) => {
+    const utcdt = new Date();
+    const localeoffset = moment.tz.zone(tz).utcOffset(utcdt.getTime());
+    const localdt = new Date(utcdt.getTime() + (-1 * localeoffset * 60 * 1000));
+    localdt.setUTCDate(localdt.getUTCDate() - localdt.getUTCDay() + (7 * offset));
 
-    return `${dt.getFullYear()}-${dt.getMonth()+1}-${dt.getDate()}`;
+    return `${localdt.getUTCFullYear()}-${localdt.getUTCMonth()+1}-${localdt.getUTCDate()}`;
 };
 
 /**
     DB登録用にデータを整形する
     params: input_data Webフォームから入力されたゴミ出しスケジュールのJSONデータ
 **/
-exports.adjustData = (input_data) => {
+exports.adjustData = (input_data, tz) => {
     let regist_data = [];
     input_data.forEach((trash)=>{
         let regist_trash = {
@@ -48,7 +48,7 @@ exports.adjustData = (input_data) => {
             if(regist_schedule.type && regist_schedule.type != 'none' && regist_schedule.value) {
                 if(regist_schedule.type === 'evweek') {
                     const offset = regist_schedule.value.start==='thisweek' ? 0 : 1;
-                    const start_date = calculateStartDate(offset);
+                    const start_date = calculateStartDate(offset, tz);
                     regist_schedule.value.start = start_date;
                 }
                 trash_schedules.push(regist_schedule);
